Create Feeforward Neural Network
Here, 10-fold cross validation is used for training and evaluation. The training fold data is read, combined, and used to train the network. The Holdout method is then used to evaluate performance. This process is repeated for about ten times to cover fold numbers 1 through 10. Resulting accuracies are obtained and averaged to yield a final estimate.

Accuracy for FNN: 54%