Create Convolutional Neural Network
To try and enhance accuracy results from the FNN, I've used a CNN.

From a singular sound clip, one can calculate log scaled mel-spectrograms and it's respective deltas. Using an input with fixed size, division/ segmentation of each sound clip into a size of 41 columns and 60 rows (60x41).

The CNN is therefore being fed two channels; mel-spec and deltas. This is elaborated on and illustrated in the Project Report for better clarity.

To prepare data for this CNN, I have used extract_feature (iterates over files within folds and appends features and class labels to arrays) and windows.

This CNN has 4 layers, proceeding with Batch Normalization and Max Pooling. Global Pooling is added after the final convolutional layer for embedding extraction, which is in turn followed by a fully-connected layer. Finally, softmax activation in a dense layer is introduced to output probabilities of audio classes. For training the network, Adam Optimizer is enforced to lower Sparse Categorical Cross-Entropy objectives.

Test sound clips undergo evaluation by averaging predictions encapsulating entire segments of audio, under the assumption that the clips are of unknown length. Finally, 10-fold accuracies are averaged to quantify network performance estimation.

Much better than our FNN result, and better than the RNN too (as you will see later), we get an accuracy of 0.6417 which is considered good, given the structure of our dataset and the intricacies behind the selection of features used for this type of classification.